solved limited portate not working as expected
A new dictionary is added to retain number or 'portate'  given away.
This  dictionary is used to solve limited 'portate' problems and now
the overselled 'portate' are reported correctly.
Learned the behavior of variable in python when we copy them !!!!!
Added some test to load_portate_ini function to avoid script abort.